+++ docs/changelog.md
@@ -3,6 +3,7 @@
 ## RELEASE
 
 * Adds `UnitySetup` [#245](https://github.com/aaronparker/evergreen/issues/245)
+* Updates `GoogleChrome` to address retirement of update source [#546](https://github.com/aaronparker/evergreen/issues/546)
 
 ## 2309.840