+++ docs/changelog.md
@@ -1,10 +1,10 @@
 # Change log
 
-## RELEASE
+## VERSION
 
 * Adds `UnitySetup` [#245](https://github.com/aaronparker/evergreen/issues/245)
 * Updates `GoogleChrome` to address retirement of update source [#546](https://github.com/aaronparker/evergreen/issues/546)
-* Updates `MicrosoftOneDrive` to add the `Throttle` property to assist filtering the current rollout version
+* Updates `MicrosoftOneDrive` to add the `Throttle` property to assist filtering the current rollout version [#547](https://github.com/aaronparker/evergreen/issues/547)
 * Removes `MicrosoftOpenJDK16`, `MicrosoftOpenJDK8` as binaries are not available
 
 ## 2309.840