+++ docs/changelog.md
@@ -1,6 +1,6 @@
 # Change log
 
-## VERSION
+## 2309.850
 
 * Adds `UnitySetup` [#245](https://github.com/aaronparker/evergreen/issues/245)
 * Updates `GoogleChrome` to address retirement of update source [#546](https://github.com/aaronparker/evergreen/issues/546)